[CSGen] forcing an optional should prefer subtyping only if it's applied to a member
`TVO_PrefersSubtypeBinding` in this case is intended to support
references to Objective-C imported members that can be force
unwrapped before being used i.e. accessing an overloaded method
on `AnyObject` should always prefer an overload that returns
a subtype:

```
@objc class A: NSObject {
    @objc func test() -> A { }
    @objc func test() -> B { }
}

func test(obj: AnyObject) {
  obj.test!()
}
```

where `B` is a subtype of `A` should result in selection of
`() -> B`.