[AST] NFC: Refactor format-specific printing into a "writer".
This makes the low-level `print*` methods in `PrintBase` simply
forward to the writer, which does the actual work. Right now,
there's only the abstract base class for the writer and the
default (S-expression-like) writer. A later commit will
introduce the JSON writer.

Also update the `printField*` methods to take `Label`s instead
of `StringRef`s.