[AST] NFC-ish: Improve printing of availability specs.
Availability specs are currently dumped by writing a pre-formatted
string directly to the output stream instead of using the structured
primitives in `PrintBase`. This needs to be fixed before we can
introduce the generalized writers.

The format here is meant to be the same as the format printed by
the original methods, but some colors may differ slightly when
dumping to the terminal.