IRGen: adjust the linkage for VWT emission on Windows
When building the standard library, we would previously mark the well
known metadata for a number of types as having external storage
(DLLImport). The well known builtin VWT are defined in the runtime
rather than the module, which would prevent us from properly annotating
the symbol. This change is an improvement but not completely sufficient
as we currently mark the VWT for `() -> ()` as having external storage.