[demangler] Add a getRemainingText() method to the demangler only for use in the debugger.
Without this one cannot easily in the debugger see what is left to be parsed
when working in a release+debuginfo build. To be careful, I only included the
method in NDEBUG builds and used SWIFT_DEBUG_HELPER so that even then, people do
not attempt to use it outside of the debugger. Finally I validated that I am not
including any additional parts of LLVM into the demangler header I am
changing (since I know the demangler is sensitive to this). The only LLVM header
I am including (llvm/Support/Compiler.h) is already included into the file by
Demangler.h... so there isnt a risk of expanding the world of LLVM being included.