TypeCheckType: Warn, not error, about missing existential `any` until Swift 7
The conditions for when and how (error or warning) a missing `any` is
diagnosed are a mouthful even if we set aside source compatibility
hacks. This change builds on
https://github.com/swiftlang/swift/pull/78347 by downgrading all missing
`any` diagnostics to warnings until Swift 7, making the behavior less
confusing for developers and giving us an opportunity to patch up
remaining holes in the diagnosis gracefully.