[sil] Add the ability to print out all types in SIL without needing to set the global flag.
This is useful if one wants to dump out state in the debugger or programatically
dump out SIL with all of the types without changing the behavior in the rest of
the compiler.

My intention is to use this to emit full types in a little bit of the logging in
SendNonSendable where seeing the full type speeds up debugging in the context of
debugging user code and where in truth, I will never ever want to suppress types
and will not use this to create test cases. It is just for quick scanning.