Ignore -enable-testing when compiling for the debugger on getEffectiveAccess
Compiling for the debugger relaxes certain restrictions, such as
accessing private members of a type. When compiling code for accessing a
member of a module built with -enable-testing, the compiler might
generate the access behind a dispatch thunk, which would be correct
for "regular" code, but not for the debugger, since dispatch thunks
don't exist for private members.

rdar://140300320