Remove -enable-ossa-modules from test/SILOptimizer/stdlib/Atomics.swift
Since Synchronization is in non-ossa, it cannot be imported into an ossa module.

We have support for recompilation of the .swiftinterface with -enable-ossa-modules under implicit modules.
However, recompilation of modules in the resource dir is disabled unless serialization status is
serialization::Status::SDKMismatch in discoverUpToDateCompiledModuleForInterface.

Removing the -enable-ossa-module flag in this test to avoid module loading failure