Don't look through ownership instructions during CSE
OSSA rauw creates copies in a lot of cases. Several of these copies
cannot be eliminated due to escapes, unimplemeneted optimization edge cases etc.

This can be an issue for performance parity with non-ossa

Disable looking through ownership instructions to avoid copies created
by OSSA rauw.

This can be revisited if we rewrite OSSA rauw to not create copies.

This has no effect on microbenchmarks with -enable-ossa-modules
on by default.

rdar://139773406