Respect "getBypassResilience" on "isResilient"
LLDB sets BypassResilience to have direct access to private properties
of modules built with resilience. This setting is ignored in many paths
that check resiliency though, which causes LLDB to fail to evaluate
certain expressions.

This commit simply makes sure that `isResilient` checks whether
resiliency should be bypassed or not.

rdar://137876089