Sema: Allow opening existential with a warning for rdar://141962317
We now correctly refuse to open the existential in the following case
due to the invariant `T` in the second parameter, which broke some code
internally. Opening the existential in this precise case happens to be
safe, so this patch adds a 6.1-only exception to the rules and pairs it
with a warning so folks can spot and fix the problem.

```swift
struct S<each T> {}
protocol P {}
func open<T: P>(_: T.Type, _: S<T>? = nil) {}
let meta: any P.Type
open(meta)
```