Workaround invalid compiler selection for LLVM Runtimes targets in cross-compilation
LLVM Runtimes contain various compiler-rt components like builtins, crt, profile and sanitizer libraries, etc.
The LLVM project spawns a nested build-tree for each configured target architecture.
For Windows we aim to build builtins, profile and sanitizer for x86_64, aarch64 and i686.

In regular builds LLVM selects the just-built Clang compiler for cross-compilation in the nested build-trees.
In cross-builds, it falls back to the default host compiler. On Windows this resolves to MSVC, which doesn't work.

We have to override the compiler in each nested build-tree manually, to select the stage-1 Clang compiler,
that we use for the LLVM Project build itself.